package/koruza/requirements.txt
requests==2.7.0
pyopenssl==0.15.1
ndg-httpsclient==0.4.0
pyasn1==0.1.8
nnpy==0.9.1
gevent==1.0.2
gevent-websocket==0.9.5
Flask==0.10.1
Flask-Webpack==0.0.7
imgurpython==1.1.6

cryptography==2.0.3
Werkzeug==0.7
Jinja2==2.4
itsdangerous==0.21
idna==2.1
greenlet==0.4.7
RPi.GPIO>=0.7.0